Moved Z stats calculation to tiled rotation loop in slow converter; fixed various compiler warnings

+++ Util.cc
@@ -12,7 +12,7 @@ std::vector<std::string> split(const std::string &str, char separator) {
 
 std::vector<hsize_t> trimAxes(const std::vector<hsize_t>& dims, int N) {
     auto trimmed = dims;
-    while (trimmed.size() > N) {
+    while (trimmed.size() > (unsigned int)N) {
         
         trimmed.erase(trimmed.begin());
     }
